abstract ModelStub
